Fix snapping
